Move `Gem.clear_paths` to a better place
The only invokation of `system_gems` that actually needs this is the one
in the global `around` hook that calls `system_gems []`. Confusingly,
the call is _not_ to make the subsequent environment modifications be
picked up inmediately by `Gem.paths`. Instead, it is to make prior
environment modifications made by the global test setup be applied, such
as the new `ENV["HOME"]` that affects `Gem.user_home` value which
`bundler` relies on.

For this reason, it is more appriopriate to make this part of the
`reset!` method that gets called only once per spec.